requirements.txt
gitpython
openai
PyYAML
Jinja2
pytest
